# docs/swagger.yaml
definitions:
  dto.ScheduleRequest:
    properties:
      class_code:
        type: string
      class_name:
        type: string
      date:
        type: string
      jam_ke:
        type: integer
      subject_code:
        type: string
      teacher_name:
        type: string
      teacher_nik:
        type: string
      time_end:
        type: string
      time_start:
        type: string
    type: object
info:
  contact: {}
  description: REST API for school schedule management.
  title: School Schedule API
  version: "1.0"
paths:
  /api/schedules:
    get:
      parameters:
      - description: Class code
        in: query
        name: class_code
        type: string
      - description: Teacher NIK
        in: query
        name: teacher_nik
        type: string
      - description: Date YYYY-MM-DD
        in: query
        name: date
        type: string
      - description: Start date YYYY-MM-DD
        in: query
        name: start_date
        type: string
      - description: End date YYYY-MM-DD
        in: query
        name: end_date
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all schedules
      tags:
      - schedules
    post:
      consumes:
      - application/json
      parameters:
      - description: Schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create schedule
      tags:
      - schedules
  /api/schedules/{id}:
    delete:
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete schedule
      tags:
      - schedules
    get:
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get schedule by ID
      tags:
      - schedules
    put:
      consumes:
      - application/json
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update schedule
      tags:
      - schedules
  /api/schedules/export:
    get:
      parameters:
      - description: Start date YYYY-MM-DD
        in: query
        name: start_date
        required: true
        type: string
      - description: End date YYYY-MM-DD
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Export JP recap to Excel
      tags:
      - excel
  /api/schedules/report/rekap-jp:
    get:
      parameters:
      - description: Start date YYYY-MM-DD
        in: query
        name: start_date
        required: true
        type: string
      - description: End date YYYY-MM-DD
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get JP recap by teacher
      tags:
      - reports
  /api/schedules/student:
    get:
      parameters:
      - description: Class code
        in: query
        name: class_code
        required: true
        type: string
      - description: Date YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get student schedule by class and date
      tags:
      - frontend
  /api/schedules/teacher:
    get:
      parameters:
      - description: Teacher NIK
        in: query
        name: teacher_nik
        required: true
        type: string
      - description: Start date YYYY-MM-DD
        in: query
        name: start_date
        required: true
        type: string
      - description: End date YYYY-MM-DD
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get teacher schedule and total JP
      tags:
      - frontend
  /api/schedules/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Excel .xlsx file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Upload schedules from Excel
      tags:
      - excel
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"